+++ src/Aerofly/Mission.ts
@@ -920,6 +920,11 @@ export class Mission {
   }
 
   toString(): string {
+    if (this.no_guides) {
+      // Create finish target plane 1m in front of aircraft origin position
+      this.finish = new MissionCheckpoint();
+      this.finish.lon_lat = this.origin_lon_lat.getRelativeCoordinates(1 / Units.meterPerNauticalMile, this.origin_dir);
+    }
     const finish = this.finish?.toStringTargetPlane("finish") ?? "";
     let string = `\
             // Exported by Aerofly Missionsgerät
@@ -943,7 +948,7 @@ ${this.conditions + finish}\
                 <[list_tmmission_checkpoint][checkpoints][]
 `;
     this.checkpoints.forEach((c, i) => {
-      string += c.toString(i, this.no_guides);
+      string += c.toString(i);
     });
     string += `\
                 >