+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## November
+
+- Fixed `no_guides` to hide guides
+- Added DR 400 aircraft
+
+## October
+
 - Added `no_guides` to hide guides
 - Added `flyOver` property to waypoints
 - Added `finish` property to missions